adds 'service optional/conditional' ability to 'Search'

+++ server/middleware/doc/Search/Update.mjs
@@ -23,7 +23,13 @@ const docSearchUpdate = (req, res, next) => {
 	)
 
 	Services.Run()
-	.then(services =>{
+	.then(services => {
+
+		if (undefined === services.sphinxql) {
+			// NON-FATAL in this middleware
+			next()
+			return
+		}
 
 		services.sphinxql.query(
 			`DELETE FROM ${CONFIG.SphinxQL.indexName} WHERE oid = ?`,
